postgres: convert postgres to datastore v2 API
