+++ internal/datastore/postgres/reader.go
@@ -1,0 +1,203 @@
+package postgres
+
+import (
+	"context"
+	"errors"
+	"fmt"
+
+	sq "github.com/Masterminds/squirrel"
+	v1 "github.com/authzed/authzed-go/proto/authzed/api/v1"
+	"github.com/jackc/pgx/v4"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/trace"
+	"google.golang.org/protobuf/proto"
+
+	"github.com/authzed/spicedb/internal/datastore/common"
+	"github.com/authzed/spicedb/internal/datastore/options"
+	"github.com/authzed/spicedb/pkg/datastore"
+	core "github.com/authzed/spicedb/pkg/proto/core/v1"
+)
+
+type pgReader struct {
+	txSource      common.TxFactory
+	querySplitter common.TupleQuerySplitter
+	filterer      queryFilterer
+}
+
+type queryFilterer func(original sq.SelectBuilder) sq.SelectBuilder
+
+var (
+	queryTuples = psql.Select(
+		colNamespace,
+		colObjectID,
+		colRelation,
+		colUsersetNamespace,
+		colUsersetObjectID,
+		colUsersetRelation,
+	).From(tableTuple)
+
+	schema = common.SchemaInformation{
+		ColNamespace:        colNamespace,
+		ColObjectID:         colObjectID,
+		ColRelation:         colRelation,
+		ColUsersetNamespace: colUsersetNamespace,
+		ColUsersetObjectID:  colUsersetObjectID,
+		ColUsersetRelation:  colUsersetRelation,
+	}
+
+	readNamespace = psql.Select(colConfig, colCreatedTxn).From(tableNamespace)
+)
+
+const (
+	errUnableToReadConfig     = "unable to read namespace config: %w"
+	errUnableToListNamespaces = "unable to list namespaces: %w"
+)
+
+func (r *pgReader) QueryRelationships(
+	ctx context.Context,
+	filter *v1.RelationshipFilter,
+	opts ...options.QueryOptionsOption,
+) (iter datastore.RelationshipIterator, err error) {
+	qBuilder := common.NewSchemaQueryFilterer(schema, r.filterer(queryTuples)).
+		FilterToResourceType(filter.ResourceType)
+
+	if filter.OptionalResourceId != "" {
+		qBuilder = qBuilder.FilterToResourceID(filter.OptionalResourceId)
+	}
+
+	if filter.OptionalRelation != "" {
+		qBuilder = qBuilder.FilterToRelation(filter.OptionalRelation)
+	}
+
+	if filter.OptionalSubjectFilter != nil {
+		qBuilder = qBuilder.FilterToSubjectFilter(filter.OptionalSubjectFilter)
+	}
+
+	return r.querySplitter.SplitAndExecuteQuery(ctx, qBuilder, opts...)
+}
+
+func (r *pgReader) ReverseQueryRelationships(
+	ctx context.Context,
+	subjectFilter *v1.SubjectFilter,
+	opts ...options.ReverseQueryOptionsOption,
+) (iter datastore.RelationshipIterator, err error) {
+	qBuilder := common.NewSchemaQueryFilterer(schema, r.filterer(queryTuples)).
+		FilterToSubjectFilter(subjectFilter)
+
+	queryOpts := options.NewReverseQueryOptionsWithOptions(opts...)
+
+	if queryOpts.ResRelation != nil {
+		qBuilder = qBuilder.
+			FilterToResourceType(queryOpts.ResRelation.Namespace).
+			FilterToRelation(queryOpts.ResRelation.Relation)
+	}
+
+	return r.querySplitter.SplitAndExecuteQuery(ctx,
+		qBuilder,
+		options.WithLimit(queryOpts.ReverseLimit),
+	)
+}
+
+func (r *pgReader) ReadNamespace(ctx context.Context, nsName string) (*core.NamespaceDefinition, datastore.Revision, error) {
+	ctx, span := tracer.Start(ctx, "ReadNamespace", trace.WithAttributes(
+		attribute.String("name", nsName),
+	))
+	defer span.End()
+
+	tx, txCleanup, err := r.txSource(ctx)
+	if err != nil {
+		return nil, datastore.NoRevision, fmt.Errorf(errUnableToReadConfig, err)
+	}
+	defer txCleanup(ctx)
+
+	loaded, version, err := loadNamespace(ctx, nsName, tx, r.filterer(readNamespace))
+	switch {
+	case errors.As(err, &datastore.ErrNamespaceNotFound{}):
+		return nil, datastore.NoRevision, err
+	case err == nil:
+		return loaded, version, nil
+	default:
+		return nil, datastore.NoRevision, fmt.Errorf(errUnableToReadConfig, err)
+	}
+}
+
+func loadNamespace(ctx context.Context, namespace string, tx pgx.Tx, baseQuery sq.SelectBuilder) (*core.NamespaceDefinition, datastore.Revision, error) {
+	ctx, span := tracer.Start(datastore.SeparateContextWithTracing(ctx), "loadNamespace")
+	defer span.End()
+
+	sql, args, err := baseQuery.Where(sq.Eq{colNamespace: namespace}).ToSql()
+	if err != nil {
+		return nil, datastore.NoRevision, err
+	}
+
+	var config []byte
+	var version datastore.Revision
+	err = tx.QueryRow(ctx, sql, args...).Scan(&config, &version)
+	if err != nil {
+		if errors.Is(err, pgx.ErrNoRows) {
+			err = datastore.NewNamespaceNotFoundErr(namespace)
+		}
+		return nil, datastore.NoRevision, err
+	}
+
+	loaded := &core.NamespaceDefinition{}
+	err = proto.Unmarshal(config, loaded)
+	if err != nil {
+		return nil, datastore.NoRevision, err
+	}
+
+	return loaded, version, nil
+}
+
+func (r *pgReader) ListNamespaces(ctx context.Context) ([]*core.NamespaceDefinition, error) {
+	ctx = datastore.SeparateContextWithTracing(ctx)
+
+	tx, txCleanup, err := r.txSource(ctx)
+	if err != nil {
+		return nil, err
+	}
+	defer txCleanup(ctx)
+
+	nsDefs, err := loadAllNamespaces(ctx, tx, r.filterer(readNamespace))
+	if err != nil {
+		return nil, fmt.Errorf(errUnableToListNamespaces, err)
+	}
+
+	return nsDefs, err
+}
+
+func loadAllNamespaces(ctx context.Context, tx pgx.Tx, query sq.SelectBuilder) ([]*core.NamespaceDefinition, error) {
+	sql, args, err := query.ToSql()
+	if err != nil {
+		return nil, err
+	}
+
+	rows, err := tx.Query(ctx, sql, args...)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+
+	var nsDefs []*core.NamespaceDefinition
+	for rows.Next() {
+		var config []byte
+		var version datastore.Revision
+		if err := rows.Scan(&config, &version); err != nil {
+			return nil, err
+		}
+
+		var loaded core.NamespaceDefinition
+		if err := proto.Unmarshal(config, &loaded); err != nil {
+			return nil, fmt.Errorf(errUnableToReadConfig, err)
+		}
+
+		nsDefs = append(nsDefs, &loaded)
+	}
+	if rows.Err() != nil {
+		return nil, rows.Err()
+	}
+
+	return nsDefs, nil
+}
+
+var _ datastore.Reader = &pgReader{}
