+++ internal/datastore/postgres/stats.go
@@ -5,8 +5,10 @@ import (
 	"fmt"
 
 	sq "github.com/Masterminds/squirrel"
+	"github.com/jackc/pgx/v4"
 
 	"github.com/authzed/spicedb/pkg/datastore"
+	corev1 "github.com/authzed/spicedb/pkg/proto/core/v1"
 )
 
 const (
@@ -27,34 +29,9 @@ var (
 )
 
 func (pgd *pgDatastore) Statistics(ctx context.Context) (datastore.Stats, error) {
-	tx, err := pgd.dbpool.Begin(ctx)
-	if err != nil {
-		return datastore.Stats{}, fmt.Errorf("unable to establish transaction: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	if pgd.analyzeBeforeStatistics {
-		if _, err := tx.Exec(ctx, fmt.Sprintf("ANALYZE %s", tableTuple)); err != nil {
-			return datastore.Stats{}, fmt.Errorf("unable to analyze tuple table: %w", err)
-		}
-	}
-
 	idSQL, idArgs, err := queryUniqueID.ToSql()
 	if err != nil {
 		return datastore.Stats{}, fmt.Errorf("unable to generate query sql: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	var uniqueID string
-	if err := tx.QueryRow(ctx, idSQL, idArgs...).Scan(&uniqueID); err != nil {
-		return datastore.Stats{}, fmt.Errorf("unable to query unique ID: %w", err)
-	}
-
-	nsQuery := readNamespace.Where(sq.Eq{colDeletedTxn: liveDeletedTxnID})
-
-	nsDefs, err := loadAllNamespaces(ctx, tx, nsQuery)
-	if err != nil {
-		return datastore.Stats{}, fmt.Errorf("unable to load namespaces: %w", err)
 	}
 
 	rowCountSQL, rowCountArgs, err := queryEstimatedRowCount.ToSql()
@@ -62,9 +39,34 @@ func (pgd *pgDatastore) Statistics(ctx context.Context) (datastore.Stats, error)
 		return datastore.Stats{}, fmt.Errorf("unable to prepare row count sql: %w", err)
 	}
 
+	nsQuery := readNamespace.Where(sq.Eq{colDeletedTxn: liveDeletedTxnID})
+
+	var uniqueID string
+	var nsDefs []*corev1.NamespaceDefinition
 	var relCount int64
-	if err := tx.QueryRow(ctx, rowCountSQL, rowCountArgs...).Scan(&relCount); err != nil {
-		return datastore.Stats{}, fmt.Errorf("unable to read relationship count: %w", err)
+	if err := pgd.dbpool.BeginTxFunc(ctx, pgd.readTxOptions, func(tx pgx.Tx) error {
+		if pgd.analyzeBeforeStatistics {
+			if _, err := tx.Exec(ctx, fmt.Sprintf("ANALYZE %s", tableTuple)); err != nil {
+				return fmt.Errorf("unable to analyze tuple table: %w", err)
+			}
+		}
+
+		if err := tx.QueryRow(ctx, idSQL, idArgs...).Scan(&uniqueID); err != nil {
+			return fmt.Errorf("unable to query unique ID: %w", err)
+		}
+
+		nsDefs, err = loadAllNamespaces(ctx, tx, nsQuery)
+		if err != nil {
+			return fmt.Errorf("unable to load namespaces: %w", err)
+		}
+
+		if err := tx.QueryRow(ctx, rowCountSQL, rowCountArgs...).Scan(&relCount); err != nil {
+			return fmt.Errorf("unable to read relationship count: %w", err)
+		}
+
+		return nil
+	}); err != nil {
+		return datastore.Stats{}, err
 	}
 
 	// Sometimes relCount can be negative on postgres, truncate to 0
